requirements.txt
openai==1.12.0
tiktoken==0.6.0
python-dotenv==1.0.1
requests==2.31.0
numpy==1.26.4
langchain-text-splitters==0.0.1
lxml==5.2.1
pandas==2.2.2